Refactor BaseCfg to BaseCfgS and update related handlers and types. Introduce ClaimDailyReward reducer and adjust dailyRewardClaimStreak type to ushort in Account. Remove obsolete BaseCfg files from project.

+++ apps/web/marbles-web-client/src/module_bindings/my_account_table.ts
@@ -19,6 +19,6 @@ export default __t.row({
   gold: __t.u32(),
   firstLoginBonusClaimed: __t.bool(),
   isMember: __t.bool(),
-  dailyRewardClaimStreak: __t.i64(),
+  dailyRewardClaimStreak: __t.u16(),
   lastDailyRewardClaimDay: __t.i64(),
 });
